backend/requirements.txt
Flask
pandas
numpy
matplotlib
seaborn
plotly
scikit-learn
Werkzeug
together
tabula-py
PyPDF2
python-docx
requests
beautifulsoup4
python-dotenv
openpyxl
lxml
flask-cors
psutil
